requirements.txt
Flask==0.12.2
requests==2.21.0
Flask-Cors==3.0.2
gunicorn==19.7.1
beautifulsoup4==4.7.1
google-api-core==1.8.0
google-auth==1.6.3
google-cloud-language==1.1.1
googleapis-common-protos==1.5.8